Move simple_clean folder inside package; move unit tests to own folder
