+++ src/dm_bip/cleaners/remove_empty_columns.py
